Got the animation for the score working finally. Thank you Framer Motion lol
