+++ client/package.json
@@ -9,11 +9,11 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "framer-motion": "^7.3.6",
     "install": "^0.13.0",
     "npm": "^8.19.2",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "styled-components": "^5.3.5"
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
     "@rollup/plugin-replace": "^4.0.0",
